+++ test/admin/fields/schemas/authentications/login.test.ts
@@ -1,4 +1,4 @@
-import loginSchema from '../../../../../src/admin/fields/schemas/authentications/loginSchema';
+import loginSchema from '../../../../../src/admin/fields/schemas/authentications/login';
 
 describe('ログインバリデーション', () => {
   test('成功', async () => {
